deadeye/config/pyproject.toml
[tool.poetry]
name = "deadeye-config"
version = "0.1.0"
description = "Download Deadeye configurations"
authors = ["Jeff Hutchison <jeff@jeffhutchison.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pynetworktables = "^2021.0.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
